@@ -2,8 +2,10 @@
 	<Import Project="..\..\Shared.props" />
 
 	<PropertyGroup>
-		<TargetFrameworks>net48;netcoreapp2.1;netcoreapp3.1;net5.0;</TargetFrameworks>
+		<TargetFrameworks>net48;netcoreapp3.1;net5.0;net6.0</TargetFrameworks>
+		<NoWarn>1591</NoWarn>
 	</PropertyGroup>
+	
 	<ItemGroup>
 		<PackageReference Include="NUnit3TestAdapter" />
 		<PackageReference Include="NUnit" />
@@ -23,4 +25,5 @@
 	    <CopyToOutputDirectory>Always</CopyToOutputDirectory>
 	  </None>
 	</ItemGroup>
+
 </Project>
