@@ -6,7 +6,7 @@ namespace SourcemapToolkit.CallstackDeminifier
 	/// <summary>
 	/// Provides factory methods to create stack trace deminifier instance.
 	/// </summary>
-	public class StackTraceDeminifierFactory
+	public sealed class StackTraceDeminifierFactory
 	{
 		private static void ValidateArguments(ISourceMapProvider sourceMapProvider, ISourceCodeProvider generatedCodeProvider, IStackTraceParser stackTraceParser)
 		{
@@ -31,10 +31,10 @@ namespace SourcemapToolkit.CallstackDeminifier
 		/// </summary>
 		/// <param name="sourceMapProvider">Consumers of the API should implement this interface, which provides the source map for a given JavaScript file. Throws ArgumentNullException if the parameter is set to null.</param>
 		/// <param name="generatedCodeProvider">Consumers of the API should implement this interface, which provides the contents of a JavaScript file. Throws ArgumentNullException if the parameter is set to null.</param>
-		public static StackTraceDeminifier GetStackTraceDeminfier(ISourceMapProvider sourceMapProvider, ISourceCodeProvider generatedCodeProvider)
-		{
-			return GetStackTraceDeminfier(sourceMapProvider, generatedCodeProvider, new StackTraceParser());
-		}
+		public static StackTraceDeminifier GetStackTraceDeminfier(
+			ISourceMapProvider sourceMapProvider,
+			ISourceCodeProvider generatedCodeProvider)
+			=> GetStackTraceDeminfier(sourceMapProvider, generatedCodeProvider, new StackTraceParser());
 
 		/// <summary>
 		/// Creates a StackTraceDeminifier with full capabilities. StackTrace deminifiers created with this method will keep source maps cached, and thus use significantly more memory during runtime than the ones generated with GetMethodNameOnlyStackTraceDeminfier.
@@ -58,10 +58,7 @@ namespace SourcemapToolkit.CallstackDeminifier
 		/// StackTrace deminifiers created with this method will keep source maps cached, and thus use significantly more memory during runtime than the ones generated with GetMethodNameOnlyStackTraceDeminfier.
 		/// </summary>
 		/// <param name="sourceMapProvider">Consumers of the API should implement this interface, which provides the source map for a given JavaScript file. Throws ArgumentNullException if the parameter is set to null.</param>
-		public static StackTraceDeminifier GetMapOnlyStackTraceDeminfier(ISourceMapProvider sourceMapProvider)
-		{
-			return GetMapOnlyStackTraceDeminfier(sourceMapProvider, new StackTraceParser());
-		}
+		public static StackTraceDeminifier GetMapOnlyStackTraceDeminfier(ISourceMapProvider sourceMapProvider) => GetMapOnlyStackTraceDeminfier(sourceMapProvider, new StackTraceParser());
 
 		/// <summary>
 		/// Creates a StackTraceDeminifier which does not depend on JS files, and is ES2015+ compatible.
@@ -92,10 +89,7 @@ namespace SourcemapToolkit.CallstackDeminifier
 		/// </summary>
 		/// <param name="sourceMapProvider">Consumers of the API should implement this interface, which provides the source map for a given JavaScript file. Throws ArgumentNullException if the parameter is set to null.</param>
 		/// <param name="generatedCodeProvider">Consumers of the API should implement this interface, which provides the contents of a JavaScript file. Throws ArgumentNullException if the parameter is set to null.</param>
-		public static StackTraceDeminifier GetMethodNameOnlyStackTraceDeminfier(ISourceMapProvider sourceMapProvider, ISourceCodeProvider generatedCodeProvider)
-		{
-			return GetMethodNameOnlyStackTraceDeminfier(sourceMapProvider, generatedCodeProvider, new StackTraceParser());
-		}
+		public static StackTraceDeminifier GetMethodNameOnlyStackTraceDeminfier(ISourceMapProvider sourceMapProvider, ISourceCodeProvider generatedCodeProvider) => GetMethodNameOnlyStackTraceDeminfier(sourceMapProvider, generatedCodeProvider, new StackTraceParser());
 
 		/// <summary>
 		/// Creates a StackTraceDeminifier that only deminifies the method names. StackTrace deminifiers created with this method will use significantly less memory during runtime than the 
