@@ -7,14 +7,16 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
 [assembly: AssemblyProduct("SourcemapTools")]
-[assembly: AssemblyCopyright("Copyright ©  2016")]
+[assembly: AssemblyCopyright("Copyright © 2021-2022")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 [assembly: ComVisible(false)]
 [assembly: Guid("00f66219-2e6c-4660-9b94-3c30e2b20729")]
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
+[assembly: AssemblyVersion("2.0.0.0")]
+[assembly: AssemblyFileVersion("2.0.0.0")]
 
 [assembly: InternalsVisibleTo("SourcemapTools.UnitTests, PublicKey=002400000480000094000000060200000024000052534131000400000100010001e16f05dfe2cb6b2f5dca681dd5b01951f45596450aea51288b7a3c44fbf8f3c0438b918a074f40a09f7743badf7421cea59dd25d41c61e96cc80c0098064ba0d7347cc53b3033d8d4063b69a6f7b03da40e3715e56ed70f4ea8b44fa70ae394fe588efdc298f59accb2f4722d22dcb806afdfbd3cca98f204fda49531a34cb")]
 // Castle proxy (used by Moq)
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2, PublicKey=0024000004800000940000000602000000240000525341310004000001000100c547cac37abd99c8db225ef2f6c8a3602f3b3606cc9891605d02baa56104f4cfc0734aa39b93bf7852f7d9266654753cc297e7d2edfe0bac1cdcf9f717241550e0a7b191195b7667bb4f64bcb8e2121380fd1d9d46ad2d92d2d15605093924cceaf74c4861eff62abf69b9291ed0a340e113be11e6a7d3113e92484cf7045cc7")]
+
+[module: SkipLocalsInit]
