@@ -1,17 +1,20 @@
-﻿namespace SourcemapToolkit.SourcemapParser
+﻿using System.Diagnostics.CodeAnalysis;
+
+namespace SourcemapToolkit.SourcemapParser
 {
 	/// <summary>
 	/// The various fields within a segment of the Mapping parser are relative to the previous value we parsed.
 	/// This class tracks this state throughout the parsing process. 
 	/// </summary>
-	internal struct MappingsParserState
+	[SuppressMessage("Performance", "CA1815:Override equals and operator equals on value types", Justification = "Not used in comparisons")]
+	internal readonly struct MappingsParserState
 	{
-		public readonly int CurrentGeneratedLineNumber;
-		public readonly int CurrentGeneratedColumnBase;
-		public readonly int SourcesListIndexBase;
-		public readonly int OriginalSourceStartingLineBase;
-		public readonly int OriginalSourceStartingColumnBase;
-		public readonly int NamesListIndexBase;
+		public int CurrentGeneratedLineNumber { get; }
+		public int CurrentGeneratedColumnBase { get; }
+		public int SourcesListIndexBase { get; }
+		public int OriginalSourceStartingLineBase { get; }
+		public int OriginalSourceStartingColumnBase { get; }
+		public int NamesListIndexBase { get; }
 
 		public MappingsParserState(MappingsParserState previousMappingsParserState = new MappingsParserState(),
 			int? newGeneratedLineNumber = null,
