@@ -21,15 +21,11 @@ namespace SourcemapToolkit.SourcemapParser
 
 			using (sourceMapStream)
 			{
-				// https://github.com/dotnet/runtime/issues/1574
-				var result = JsonSerializer.DeserializeAsync<SourceMap>(sourceMapStream).AsTask().GetAwaiter().GetResult();
+				var result = JsonSerializer.Deserialize<SourceMap>(sourceMapStream);
 				if (result != null)
 				{
 					// Since SourceMap is immutable we need to allocate a new one and copy over all the information
 					var parsedMappings = MappingsListParser.ParseMappings(result.Mappings ?? string.Empty, result.Names ?? new List<string>(), result.Sources ?? new List<string>());
-
-					// Resize to free unused memory
-					parsedMappings.Capacity = parsedMappings.Count;
 
 					result = new SourceMap(
 						result.Version,
